+++ tools/scripts/nuget-clear-local.sh
@@ -35,7 +35,7 @@ done
 
 # Clear the local NuGet feed folder
 echo "Clearing local NuGet feed folder: $local_feed_dir"
-rm -rf "$local_feed_dir"/* &>/dev/null
+rm -rf "${local_feed_dir:?}"/* &>/dev/null
 
 if [[ $? != 0 ]]; then
     echo "WARNING: Failed to clear local NuGet feed folder: $local_feed_dir"
